feat: Define protocol fee as % of notional, similar to factoring pool (#70)
* feat: Define protocol fee as % of notional, similar to factoring pool

* feat: add total gross interest paid

+++ test/foundry/BullaFrendLend/BatchFunctionality.t.sol
@@ -38,7 +38,7 @@ contract TestBullaFrendLendBatchFunctionality is BullaFrendLendTestHelper {
     ERC20PermitMock public permitToken;
 
     uint256 constant FEE = 0.01 ether;
-    uint256 constant PROTOCOL_FEE_BPS = 1000; // 10%
+    uint16 constant PROTOCOL_FEE_BPS = 1000; // 10%
 
     function setUp() public {
         weth = new WETH();
